feat(structure): split up auth functions
